refactor(front): sépare le hook d'authentification de son fournisseur (#33)
