+++ frontend/src/pages/Inscription.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Link, Navigate, useNavigate } from 'react-router-dom';
-import { useAuthentification } from '../contexte/Authentification';
+import { useAuthentification } from '../contexte/contexteAuthentification';
 import Bouton from '../composants/Bouton';
 import Champ from '../composants/Champ';
 import Alerte from '../composants/Alerte';
