+++ frontend/src/pages/__tests__/Connexion.test.jsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, afterEach } from 'vitest';
 import { render, screen, cleanup, waitFor } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import Connexion from '../Connexion';
-import * as contexte from '../../contexte/Authentification';
+import * as contexte from '../../contexte/contexteAuthentification';
 
 afterEach(() => {
   cleanup();
